Refactor project structure and update secrets handling
Updated project files to include a new shared project reference to `Mattermost.Models`. Refactored `Program.cs` to read secrets from `secrets.json` and use them in the `LoginAsync` method. Moved `Secrets` class to `Mattermost.Models` namespace and updated its definition. Adjusted `MattermostClientTests.cs` to reflect these changes. Updated solution file to include the new shared project.
