+++ Sources/Mattermost.Models/Secrets.cs
@@ -1,0 +1,16 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Mattermost.Models
+{
+    public class Secrets
+    {
+        [JsonPropertyName("username")]
+        public string Username { get; set; } = string.Empty;
+
+        [JsonPropertyName("password")]
+        public string Password { get; set; } = string.Empty;
+
+        [JsonPropertyName("token")]
+        public string Token { get; set; } = string.Empty;
+    }
+}
