+++ Sources/Mattermost.sln
@@ -9,6 +9,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mattermost", "Mattermost\Ma
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mattermost.Tests", "Mattermost.Tests\Mattermost.Tests.csproj", "{67CAFAB0-3116-4947-B9B2-A4CFDACE6287}"
 EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Mattermost.Models", "Mattermost.Models\Mattermost.Models.shproj", "{D0DC5188-CC72-45AE-B3D4-D235F3546FD9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,5 +35,10 @@ Global
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {326BD828-2995-4180-8201-3CE9CA8F92C5}
+	EndGlobalSection
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		Mattermost.Models\Mattermost.Models.projitems*{67cafab0-3116-4947-b9b2-a4cfdace6287}*SharedItemsImports = 5
+		Mattermost.Models\Mattermost.Models.projitems*{78fbf3fc-d72d-45de-bdb3-c147346ad7d6}*SharedItemsImports = 5
+		Mattermost.Models\Mattermost.Models.projitems*{d0dc5188-cc72-45ae-b3d4-d235f3546fd9}*SharedItemsImports = 13
 	EndGlobalSection
 EndGlobal
